fix overflows on req arg in ioctl

+++ sys/ioctl.go
@@ -9,9 +9,9 @@ import (
 )
 
 // Ioctl wraps the ioctl syscall.
-func Ioctl(fd int, req int, arg uintptr) error {
+func Ioctl(fd int, req uintptr, arg uintptr) error {
 	_, _, e := unix.Syscall(
-		unix.SYS_IOCTL, uintptr(fd), uintptr(req), arg)
+		unix.SYS_IOCTL, uintptr(fd), req, arg)
 	if e != 0 {
 		return os.NewSyscallError("ioctl", e)
 	}
